feat: :sparkles: style table using css grid to allow scrollable tbody
